backend/apps/users/api/serializers/users.py
from rest_framework import serializers
from users.models import User


class SignUpSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ("email", "firebase_id")
