backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Roman Boba <9119android9119@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
black = "^23.7.0"
django-silk = "^5.0.3"
isort = "^5.12.0"
flake8 = "^6.1.0"
firebase = "^4.0.1"
firebase-admin = "^6.2.0"
psycopg2 = "^2.9.7"
django-cors-headers = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
